Resolve #38 and remove convolution
This offloads a lot of WTabHolder's work to WPanel by having each tab store its widgets in a WPanel. This removes a lot of convolution by only requiring interactions with one object and removes some code that would be duplicated in WPanel. It also resolves #38 by allowing access to the body panel.